+++ c_impl/include/filter_analyzer.h
@@ -12,20 +12,23 @@
  * processing HOP_SAMPLES samples per call. The HPF is the AEC3 verbatim 3-tap
  * minimum-phase ~600 Hz cutoff filter.
  *
- * BYTE-EQUAL parity notes (numpy 1.26 -> C, captured from real balanced/DT
- * case wav/aec_challenge_blind/doubletalk/0I0XMl3M..._doubletalk_*.wav,
- * preset=balanced fl=832 hop=160):
+ * FLOAT32-BY-DESIGN notes (originally captured from numpy 1.26 -> C parity on
+ * a real balanced/DT case wav/aec_challenge_blind/doubletalk/0I0XMl3M...
+ * _doubletalk_*.wav, preset=balanced fl=832 hop=160; the module now runs
+ * entirely in float32 — the former fp64 accumulation paths below were
+ * converted for uniformity as part of the f32 campaign, drift accepted):
  *   - filter_taps : float32, size 960 (6 partitions * 160).
  *   - render_block: float32, size 160 (= render_block_scaled, int16 amplitude).
  *   - _h_highpass : float32. HPF taps (float32) * filter_taps (float32) all f32.
- *   - _gain : python float (f64); _peak_index / _delay_blocks : python int.
+ *   - _gain : float32; _peak_index / _delay_blocks : int.
  *   - ConsistentFilterDetector floor accumulators:
  *       * np.abs(filter_taps_slice).sum()  -> float32 PAIRWISE sum
- *         (f32_pairwise_sum), then float()->f64 accumulate into _floor_accum.
- *       * seg.max(initial=0.0)             -> float32 max, widened to f64.
+ *         (f32_pairwise_sum), accumulated directly into _floor_accum (float32).
+ *       * seg.max(initial=0.0)             -> float32 max, stored directly.
  *       * render active power
- *         (render_block.astype(f64)**2).sum() -> float64 PAIRWISE sum
- *         (f64_pairwise_sum) compared > _active_render_threshold (f64).
+ *         (render_block**2).sum() -> float32 PAIRWISE sum
+ *         (f32_pairwise_sum, formerly f64_pairwise_sum) compared >
+ *         _active_render_threshold (float32).
  *   - peak find: h[k]*h[k] in float32; np.argmax returns FIRST max index.
  *
  * Build with -ffp-contract=off so the float32 multiply/add/compare is not fused.
@@ -45,10 +48,10 @@
 
 /* ConsistentFilterDetector mirror state (reset each full-filter sweep). */
 typedef struct {
-    double active_render_threshold;  /* (active_render_limit^2) * HOP_SAMPLES   */
+    float  active_render_threshold;  /* (active_render_limit^2) * HOP_SAMPLES   */
     int    significant_peak;         /* bool                                    */
-    double floor_accum;              /* f64                                     */
-    double secondary_peak;           /* f64 (holds widened f32 max)             */
+    float  floor_accum;              /* float32                                 */
+    float  secondary_peak;           /* float32 (holds the f32 max)             */
     int    floor_low_limit;
     int    floor_high_limit;
     long   counter;
@@ -56,16 +59,16 @@ typedef struct {
 } FaConsistentDetector;
 
 typedef struct {
-    double active_render_threshold;  /* (active_render_limit^2) * HOP_SAMPLES   */
+    float  active_render_threshold;  /* (active_render_limit^2) * HOP_SAMPLES   */
     int    bounded_erl;              /* bool                                    */
-    double default_gain;
+    float  default_gain;
     FaConsistentDetector consistent;
 
     int    region_start;
     int    region_end;
     long   blocks_since_reset;
     int    peak_index;
-    double gain;
+    float  gain;
     int    consistent_estimate;      /* bool                                    */
     int    delay_blocks;
 
@@ -77,7 +80,7 @@ typedef struct {
  * active_render_limit / bounded_erl / default_gain match the Python defaults
  * (100.0 / 0 / 1.0) unless overridden. `size` is the full filter length. */
 void fa_init(FilterAnalyzer *m, float *h_highpass_storage, int size,
-             double active_render_limit, int bounded_erl, double default_gain);
+             float active_render_limit, int bounded_erl, float default_gain);
 
 void fa_reset(FilterAnalyzer *m);
 
@@ -89,12 +92,15 @@ void fa_update(FilterAnalyzer *m, const float *filter_taps,
 /* Public queries. */
 int    fa_min_filter_delay_blocks(const FilterAnalyzer *m);
 int    fa_any_filter_consistent(const FilterAnalyzer *m);
-double fa_max_echo_path_gain(const FilterAnalyzer *m);
+float  fa_max_echo_path_gain(const FilterAnalyzer *m);
 int    fa_peak_index(const FilterAnalyzer *m);
 const float *fa_get_adjusted_filters(const FilterAnalyzer *m);
 
-/* numpy-1.26-bit-exact pairwise sums (exposed for the parity test). */
+/* float32 pairwise sums (numpy-1.26-shaped reduction tree; exposed for the
+ * parity test). fa_f64_pairwise_sum is float32-by-design now (formerly a
+ * float64 accumulator variant used for the render active-power sum) — kept
+ * as a distinct entry point since callers still reference it by name. */
 float  fa_f32_pairwise_sum(const float *a, size_t n);
-double fa_f64_pairwise_sum(const double *a, size_t n);
+float  fa_f64_pairwise_sum(const float *a, size_t n);
 
 #endif /* FILTER_ANALYZER_H */
